create LogAnalyzer
기존 LogAnalyzer는 LogAnalyzercmd로 변경 리팩토링 후 LogAnalyzercmd는 삭제할 예정
