+++ src/main/java/LogFileReader.java
@@ -5,9 +5,10 @@ import java.io.IOException;
 
 public class LogFileReader {
     private static final String OUTPUT_FILE_NAME = "output.log";
-    private BufferedReader br = new BufferedReader(new FileReader(OUTPUT_FILE_NAME));
+    private BufferedReader br;
 
     public LogFileReader() throws FileNotFoundException {
+        this.br = new BufferedReader(new FileReader(OUTPUT_FILE_NAME));
     }
 
     public String read() throws IOException {
